backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
pydantic[email]==2.7.1
python-dotenv==1.0.1
email-validator==2.1.1
Pillow==10.3.0
aiofiles==23.2.1
